feat[Yahoo_DSP]: Refactor Code
