+++ v0/destinations/yahoo_dsp/util.js
@@ -1,7 +1,8 @@
 const qs = require("qs");
+const sha256 = require("sha256");
 const { generateJWTToken } = require("../../../util/jwtTokenGenerator");
-const { httpSend, httpPOST } = require("../../../adapters/network");
-const { CustomError } = require("../../util");
+const {  httpPOST } = require("../../../adapters/network");
+const { CustomError, isDefinedAndNotNullAndNotEmpty } = require("../../util");
 
 const { ACCESS_TOKEN_CACHE_TTL } = require("./config.js");
 const Cache = require("../../util/cache");
@@ -14,8 +15,8 @@ const getUnixTimestamp = () => {
 
 /**
  * There are four location types which are stored in the poiLocationType error. These are ["chains", "woeids", "gids", "categories"].
- * The ones to be added in includes Object will be taken from audienceType with include as Prefix. The ones to be added in excludes 
- * Object will be taken from audienceType with exclude as Prefix.   
+ * The ones to be added in includes Object will be taken from audienceType with include as Prefix. The ones to be added in excludes
+ * Object will be taken from audienceType with exclude as Prefix.
  */
 const poiLocationType = [
   "includeChains",
@@ -30,8 +31,81 @@ const poiLocationType = [
 let listType;
 const includes = {};
 const excludes = {};
+let seedList = [];
+let outputPayload = {};
 
-// updating includes 
+
+/**
+ *
+ * @param {*} attributeArray  - It contains the audience lists to be added in the form of array". eg.
+ * [{"email": "abc@email.com"},{"email": "abc@email.com"},{"email": "abc@email.com"}]
+ * @param {*} Config
+ * @returns The function returns an array of Audience List provided by the user like "email", "deviceId", "ipAddress".
+ * eg. [
+ * "251014dafc651f68edac7",
+ * "afbc34416ac6e7fbb9734",
+ * "42cbe7eebb412bbcd5b56",
+ * "379b4653a40878da7a584"
+ * ]
+ */
+const populateIdentifiers = (attributeArray, { Config }) => {
+  const seedList = [];
+  const { audienceType } = Config;
+  const { hashRequired } = Config;
+  let listType;
+  if (isDefinedAndNotNullAndNotEmpty(attributeArray)) {
+    // traversing through every element in the add array for the elements to be added.
+    attributeArray.forEach(element => {
+      // storing keys of an object inside the add array.
+      const keys = Object.keys(element);
+      // checking for the audience type the user wants to add is present in the input or not.
+      keys.forEach(key => {
+        if (key === audienceType) {
+          listType = audienceType;
+        }
+      });
+      // throwing error if the audience type the user wants to add is not present in the input.
+      if (!listType) {
+        throw new CustomError(
+          `[Yahoo_DSP]:: Required property for ${audienceType} type audience is not available in an object`,
+          400
+        );
+      }
+      // here, hashing the data if is not hashed and pushing in the seedList array.
+      if (hashRequired) {
+        seedList.push(sha256(element[audienceType]));
+      } else {
+        seedList.push(element[audienceType]);
+      }
+    });
+  }
+  return seedList;
+};
+
+/**
+ * This function is used to create the output Payload.
+ * @param {*} audienceList - This is the list of audiences (in the form of a bunch of objects inside an array) to be updated. eg. [{},{},{}]
+ * @param {*} destination 
+ * @returns 
+ */
+const createPayload = (audienceList, destination) => {
+  const accountId = destination.Config.accountId;
+  let seedList = [];
+   // Populating Seed List that conains audience list to be updated
+  seedList = populateIdentifiers(audienceList, destination);
+  // throwing the error if nothing is present in the seedList
+  if (seedList.length === 0) {
+    throw new CustomError(
+      `[Yahoo_DSP]:: No attributes are present in the '${key}' property.`,
+      400
+    );
+  }
+  // Creating outputPayload
+  outputPayload = { ...outputPayload, accountId, seedList };
+  return outputPayload;
+};
+
+// updating includes
 const populateIncludes = (audienceList, audienceType) => {
   audienceList.forEach(element => {
     const audieceListkeys = Object.keys(element);
@@ -41,7 +115,10 @@ const populateIncludes = (audienceList, audienceType) => {
       }
     });
     if (!listType) {
-      throw new CustomError(`[Yahoo_DSP]:: Required property for ${audienceType} type audience is not available in an object`, 400);
+      throw new CustomError(
+        `[Yahoo_DSP]:: Required property for ${audienceType} type audience is not available in an object`,
+        400
+      );
     }
     if (element.includeChains) {
       if (!includes.chains) {
@@ -81,7 +158,10 @@ const populateExcludes = (audienceList, audienceType) => {
       }
     });
     if (!listType) {
-      throw new CustomError(`[Yahoo_DSP]:: Required property for ${audienceType} type audience is not available in an object`, 400);
+      throw new CustomError(
+        `[Yahoo_DSP]:: Required property for ${audienceType} type audience is not available in an object`,
+        400
+      );
     }
     if (element.excludeChains) {
       if (!excludes.chains) {
@@ -114,11 +194,11 @@ const populateExcludes = (audienceList, audienceType) => {
 /**
  * The funciton here is used to generate acccess token using POST call which needs some parameters like clientId, clientSecret which is being
  * taken from destination.Config and JWT token (generated using jwtTokenGenerator).
- * @param {*} destination 
- * @returns 
+ * @param {*} destination
+ * @returns
  */
 const getAccessToken = async destination => {
-  const {clientId, clientSecret} = destination.Config;
+  const { clientId, clientSecret } = destination.Config;
   const accessTokenKey = destination.ID;
 
   /**
@@ -152,22 +232,26 @@ const getAccessToken = async destination => {
         realm: "dsp",
         client_assertion_type:
           "urn:ietf:params:oauth:client-assertion-type:jwt-bearer",
-           // Here, generateJWTToken is used to get JWT required for genrating access token.
+        // Here, generateJWTToken is used to get JWT required for genrating access token.
         client_assertion: generateJWTToken(header, data, secret)
       }),
       method: "POST"
     };
     const response = await httpPOST(request.url, request.data, request.header);
     // If the request fails, throwing error.
-    if(response.success === false){
-      throw new CustomError(`[Yahoo_DSP]:: access token could not be gnerated due to ${response.response.response.data.error}`,400);
+    if (response.success === false) {
+      throw new CustomError(
+        `[Yahoo_DSP]:: access token could not be gnerated due to ${response.response.response.data.error}`,
+        400
+      );
     }
-    return response.response?.data?.access_token
+    return response.response?.data?.access_token;
   });
 };
 
 module.exports = {
   getAccessToken,
   populateIncludes,
-  populateExcludes
+  populateExcludes,
+  createPayload
 };
